Move and extract constants
